requirements.txt
fastapi~=0.115.2
uvicorn~=0.32.0
PyPDF2~=3.0.1
python-multipart~=0.0.12
python-jose~=3.3.0
passlib~=1.7.4
python-dotenv~=0.21.1
h11~=0.14.0
pip~=24.0
rsa~=4.9
pyasn1~=0.6.1
ecdsa~=0.19.0
sniffio~=1.3.1
typing_extensions~=4.12.2
six~=1.16.0
click~=8.1.7
certifi~=2025.1.31
starlette~=0.40.0
pydantic~=2.9.2
pydantic_core~=2.23.4
requests~=2.30.0
anyio~=4.6.2.post1
urllib3~=2.3.0
annotated-types~=0.7.0
idna~=3.10
charset-normalizer~=3.4.1